fix(review): consolidate fleet test helpers, add state file guard, strengthen error assertions
- Move _simple_prompt_builder, _no_sleep_quota_checker, _noop_quota_refresher
  to tests/fleet/_helpers.py (were duplicated across 5 fleet test files)
- Add len(state_files)==1 assertion to _read_state_file for diagnostic clarity
- Assert specific error code and campaign ref in user_visible_message for
  test_missing_campaign_ref_returns_fleet_error
- Assert campaign.research_dir in user_visible_message for
  test_partial_capture_propagates_only_captured_keys
